MB-9159 Batch read uses a key instead of a sequence number.
When an item is evicted from memory, we plan to evict all the
metadata fields including the item's sequence number in addiiton
to the item's value. This new eviction approach requires a
batch read to use a key instead of a sequence number when it
reads items from disk.

This change adapts the batch read, so that it can use a key to
fetch an item's metadata or value from disk.

Change-Id: Ief14fd0ef4f42b316e75c6719349cdb1c63fac80