CBD-1034 Support the full eviction as an optional eviction policy
When an item is evicted from the cache, its value is only
evicted from memory by default. To further reduce the memory
overhead, this change supports the full eviction as an additional
option, which removes an item's key, metadata, and value from the
cache. The eviction policy can be chosen when a bucket is
instantiated.

If the full eviction option is chosen, this results in affecting
lots of engine APIs and also causes more disk IO overhead.
For example, CAS for a non-resident item will require a backgroud
fetch because it should know the item's CAS value.

Change-Id: I2edbe2b12527ef1c1fb31992cb04afbb7c078dac