CBD-1040 Adapt warmup to not load all the keys with full eviction
When the full eviction is selected, we don't need to load all the
keys and their metadata during warmup. This means that the warmup
process skips "key_dump" phase and enters into
"loading_access_log" phase.

If the access log is corrupted or not found, then the warmup
process will enter into "loading_kv_pairs" phase.

Change-Id: Ia2a04e18d4f051ba5aac0a6b26d8e8f3073ca83c