MB-11059: Support datatype during offline upgrade
- When a 3.0 server reads from a pre-3.0 couchstore
database, we need to identify the value to set the
correct datatype. Warmup will handle this.

- In the loadKVPairs and loadData phase, if the
metadata buffer's length is found to be 16B, we will
check for the JSONness of the value to set the correct
datatype for the particular item in memory.

- Access log loading wouldn't even run in this scenario
as the accesslog path differs in 3.0 and pre-3.0.

- In a BGFetch case, if the metadata buffer length of
the item is found to be 16B, we identify the data for
key, before setting its datatype.

- During compaction, if the metadata buffer length is
found to be 16B, we should realloc this buffer length to
the new length (considering an additional 2B), and set
the datatype correctly by identifying the value content.

Change-Id: I1149b937d495d5dd1f0d344f701565ba96fe8225